i2c驱动抽象
