+++ components/i2c_device/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(COMPONENT_SRCS "i2c_device.c")
+set(COMPONENT_ADD_INCLUDEDIRS ".")
+set(COMPONENT_PRIV_REQUIRES "driver")
+register_component()
