+++ components/motor_dev/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(COMPONENT_SRCS "motor2204.c")
+set(COMPONENT_ADD_INCLUDEDIRS ".")
+set(COMPONENT_PRIV_REQUIRES "driver" "esp_timer")
+register_component()
